+++ source/common/types/environment.types.js
@@ -27,7 +27,6 @@ export type Environment = {
   isWindows: boolean,
   isMacOS: boolean,
   isLinux: boolean,
-  stateDirectoryPath: string,
   isInSafeMode: boolean,
 };